requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
click==7.1.2
Django==3.1.1
django-celery-results==1.2.1
gunicorn==20.0.4
h11==0.10.0
kombu==4.6.11
psycopg2==2.8.6
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
uvicorn==0.12.0
vine==1.3.0
django-ckeditor==6.0.0
